+++ PIR/include/passes/test.h
@@ -12,31 +12,16 @@
 
 #include <Evo.h>
 
-namespace pcit::core{
+#include <PCIT_core.h>
 
-	struct Version{
-		uint16_t major;
-		uint16_t release;
-		uint16_t minor;
-		uint16_t patch;
-	};
+#include "../PassManager.h"
 
-	constexpr auto version = Version{
-		.major   = 0,
-		.release = 0,
-		.minor   = 60,
-		.patch   = 0,
-	};
+namespace pcit::pir::passes{
+
+
+	EVO_NODISCARD auto test() -> PassManager::ReverseStmtPassGroup;
+
 
 }
- 	
 
-template<>
-struct std::formatter<pcit::core::Version> : std::formatter<std::string> {
-    auto format(const pcit::core::Version& version, std::format_context& ctx) const -> std::format_context::iterator {
-        return std::formatter<std::string>::format(
-        	std::format("{}.{}.{}.{}", version.major, version.release, version.minor, version.patch),
-        	ctx
-        );
-    }
-};
+