+++ Panther/exec/main.cpp
@@ -99,15 +99,15 @@ auto main(int argc, const char* argv[]) -> int {
 		pthr::printTitle(printer);
 
 		#if defined(PCIT_BUILD_DEBUG)
-			printer.printlnMagenta(std::format("v{} (debug)", pcit::core::version));
+			printer.printlnMagenta("v{} (debug)", pcit::core::version);
 		#elif defined(PCIT_BUILD_DEV)
-			printer.printlnMagenta(std::format("v{} (dev)", pcit::core::version));
+			printer.printlnMagenta("v{} (dev)", pcit::core::version);
 		#elif defined(PCIT_BUILD_OPTIMIZE)
-			printer.printlnMagenta(std::format("v{} (optimize)", pcit::core::version));
+			printer.printlnMagenta("v{} (optimize)", pcit::core::version);
 		#elif defined(PCIT_BUILD_RELEASE)
-			printer.printlnMagenta(std::format("v{} (release)", pcit::core::version));
+			printer.printlnMagenta("v{} (release)", pcit::core::version);
 		#elif defined(PCIT_BUILD_DIST)
-			printer.printlnMagenta(std::format("v{}", pcit::core::version));
+			printer.printlnMagenta("v{}", pcit::core::version);
 		#else
 			#error Unknown or unsupported build
 		#endif